move suggestion utils into discord utils instead
