+++ src/main/java/net/hypixel/nerdbot/listener/SuggestionListener.java
@@ -31,7 +31,6 @@ import net.hypixel.nerdbot.metrics.PrometheusMetrics;
 import net.hypixel.nerdbot.repository.DiscordUserRepository;
 import net.hypixel.nerdbot.repository.GreenlitMessageRepository;
 import net.hypixel.nerdbot.util.DiscordUtils;
-import net.hypixel.nerdbot.util.SuggestionUtils;
 import net.hypixel.nerdbot.util.ArrayUtils;
 import org.jetbrains.annotations.NotNull;
 
@@ -172,7 +171,7 @@ public class SuggestionListener {
     @SubscribeEvent
     public void onChannelDelete(ChannelDeleteEvent event) {
         if (event.getChannelType() == net.dv8tion.jda.api.entities.channel.ChannelType.FORUM) {
-            Suggestion.ChannelType channelType = SuggestionUtils.getForumSuggestionType(event.getChannel().asForumChannel());
+            Suggestion.ChannelType channelType = DiscordUtils.getForumSuggestionType(event.getChannel().asForumChannel());
 
             if (channelType == Suggestion.ChannelType.ALPHA || channelType == Suggestion.ChannelType.PROJECT) {
                 BotConfig botConfig = NerdBotApp.getBot().getConfig();
@@ -188,7 +187,7 @@ public class SuggestionListener {
 
     private void updateConfigForumIds(GenericChannelEvent event) {
         if (event.getChannelType() == net.dv8tion.jda.api.entities.channel.ChannelType.FORUM) {
-            Suggestion.ChannelType channelType = SuggestionUtils.getForumSuggestionType(event.getChannel().asForumChannel());
+            Suggestion.ChannelType channelType = DiscordUtils.getForumSuggestionType(event.getChannel().asForumChannel());
 
             if (channelType == Suggestion.ChannelType.ALPHA || channelType == Suggestion.ChannelType.PROJECT) {
                 BotConfig botConfig = NerdBotApp.getBot().getConfig();
